pyproject.toml
[project]
name = "ashdocextractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mistralai>=1.9.11",
    "openai>=2.8.0",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "streamlit>=1.51.0",
    "tenacity>=9.1.2",
]
